refactor(#443): share REJECTION_VERBS + classifier scaffold (simplify round)
Reuse angle: OutputModeProber and ReasoningStripProber each defined
their own REJECTION_VERBS constant and two-marker classifier (verb +
field-marker substring AND-pattern). The constants had drifted
(output-mode added 'must be' / 'should be' that reasoning-strip
lacked). Hand-keeping them in sync is a foot-gun.

Extract:

- src/llm-sdk/shared-rejection-verbs.ts (NEW): REJECTION_VERBS union
  (8 verbs from both probes — adding more only widens matches, so the
  reasoning-strip classifier behavior is monotonically broadened with
  no semantic regression) + classifyFieldError(body, fieldMarkers)
  helper.

Wire both classifiers through the helper:

- OutputModeProber.isJsonSchemaFieldError / isJsonObjectFieldError →
  classifyFieldError(body, JSON_SCHEMA_FIELD_MARKERS /
  JSON_OBJECT_FIELD_MARKERS)
- ReasoningStripProber.isReasoningFieldError →
  classifyFieldError(body, FIELD_MARKERS)

307/307 tests pass; LM Studio regression guard preserved (the
classifier input contract — responseBody not message — is now
documented in one place: shared-rejection-verbs.ts).
