+++ src/llm-sdk/reasoning-strip-probe.ts
@@ -3,6 +3,7 @@
 // v1.26.0 Batch 6: Layer-3 of the 4-layer fallback for force-disable thinking.
 //
 // Some openai-compat backends (notably Gemini-via-OpenAI-shim, Issue #137)
+import { classifyFieldError } from './shared-rejection-verbs';
 // reject `reasoning_effort: 'none'` with HTTP 400. We catch the 400 by
 // inspecting the error message for a *rejection verb* AND a *field marker*
 // (two-marker pattern, mirrors the established
@@ -48,19 +49,14 @@
 //   granularity would over-invalidate the cache.
 
 /**
- * Rejection verbs — what a backend says when it does NOT accept a field.
- * Single-substring on the lowercased error message. Chosen to match the
- * patterns an HTTP 400 with a JSON body typically uses. "Unrecognized" /
- * "unknown" / "invalid value" are the most common.
+ * Rejection verbs shared with OutputModeProber via
+ * src/llm-sdk/shared-rejection-verbs.ts (v1.26.3 PATCH simplify
+ * round). Originally 6 verbs in this file; the shared constant adds
+ * 'must be' / 'should be' which only widen matches — every string
+ * the previous local list matched still matches. The reasoning
+ * classifier has never asserted a max-set invariant, so broadening
+ * is safe.
  */
-const REJECTION_VERBS = [
-  'unrecognized',
-  'unknown',
-  'invalid value',
-  'unsupported',
-  'not allowed',
-  'not supported',
-] as const;
 
 /**
  * Field markers — names of the reasoning-related fields, as they would
@@ -152,9 +148,6 @@ export class ReasoningStripProber {
    * pass `err.responseBody`.
    */
   static isReasoningFieldError(body: string): boolean {
-    const lower = body.toLowerCase();
-    const hasVerb = REJECTION_VERBS.some((v) => lower.includes(v));
-    const hasField = FIELD_MARKERS.some((f) => lower.includes(f));
-    return hasVerb && hasField;
+    return classifyFieldError(body, FIELD_MARKERS);
   }
 }
